Share CLI defaults with GUI and expand tests
